test: move classes to appropriate packages (entity, dto, repository, etc.)
