+++ src/test/kotlin/ecommerce/controller/GetProductControllerTest.kt
@@ -1,9 +1,9 @@
-package ecommerce.product.controller
+package ecommerce.controller
 
-import ecommerce.product.data.ProductResponse
-import ecommerce.product.helper.ProductTestFixture.ValidRequest.AMERICANO
-import ecommerce.product.helper.ProductTestFixture.ValidRequest.FLAT_WHITE
-import ecommerce.product.helper.ProductTestFixture.postTestProduct
+import ecommerce.dto.ProductResponse
+import ecommerce.helper.ProductTestFixture.ValidRequest.AMERICANO
+import ecommerce.helper.ProductTestFixture.ValidRequest.FLAT_WHITE
+import ecommerce.helper.ProductTestFixture.postTestProduct
 import io.restassured.RestAssured
 import io.restassured.common.mapper.TypeRef
 import io.restassured.http.ContentType
