+++ src/test/kotlin/ecommerce/helper/MemberTestFixture.kt
@@ -1,6 +1,6 @@
-package ecommerce.auth.helper
+package ecommerce.helper
 
-import ecommerce.auth.data.MemberRequest
+import ecommerce.dto.MemberRequest
 import io.jsonwebtoken.Jwts
 
 object MemberTestFixture {
